+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - ReleaseDate
+### Added
+- Derived `Default` for `Mutex` and `RwLock`
+  ([#22](https://github.com/asomers/futures-locks/pull/22))
+  
 ## [0.4.0] - 2019-08-24
 ### Added
 - Added `Mutex::ptr_eq()`